backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
pandas>=2.1.0
python-dateutil==2.8.2
numpy>=1.24.0